# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Armor API
  version: 0.1.0
  description: >
    Agent Armor is an open-source governance layer for AI agents. It intercepts
    tool calls, evaluates risk, enforces workspace policies, and provides a
    full audit trail across an 8-layer security stack.
  license:
    name: BUSL-1.1
    url: https://spdx.org/licenses/BUSL-1.1.html
  contact:
    name: Agent Armor Team
    url: https://github.com/agent-armor/agent-armor
servers:
  - url: http://localhost:4010
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /v1/inspect:
    post:
      operationId: inspectAction
      summary: Submit an agent action for governance evaluation
      tags: [Core]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectRequest"
      responses:
        "200":
          description: Governance result with decision, risk, audit trail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/audit:
    get:
      operationId: listAuditEvents
      summary: List recent audit events
      tags: [Audit]
      responses:
        "200":
          description: Array of stored audit events (most recent 100)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoredAuditEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/reviews:
    get:
      operationId: listReviews
      summary: List all pending and resolved reviews
      tags: [Reviews]
      responses:
        "200":
          description: Array of review requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/reviews/{id}:
    post:
      operationId: resolveReview
      summary: Approve or reject a pending review
      tags: [Reviews]
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewBody"
      responses:
        "200":
          description: Updated review request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/profiles:
    get:
      operationId: listProfiles
      summary: List all agent profiles
      tags: [Profiles]
      responses:
        "200":
          description: Array of agent profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createProfile
      summary: Create or update an agent profile
      tags: [Profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentProfile"
      responses:
        "200":
          description: Upserted agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/profiles/{agent_id}:
    get:
      operationId: getProfile
      summary: Get an agent profile by ID
      tags: [Profiles]
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateProfile
      summary: Update an agent profile
      tags: [Profiles]
      parameters:
        - $ref: "#/components/parameters/AgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentProfile"
      responses:
        "200":
          description: Updated agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteProfile
      summary: Delete an agent profile
      tags: [Profiles]
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "204":
          description: Profile deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/workspaces:
    get:
      operationId: listWorkspaces
      summary: List all workspace policies
      tags: [Workspaces]
      responses:
        "200":
          description: Array of workspace policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspacePolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createWorkspace
      summary: Create or update a workspace policy
      tags: [Workspaces]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspacePolicy"
      responses:
        "200":
          description: Upserted workspace policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspacePolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      summary: Get a workspace policy by ID
      tags: [Workspaces]
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspacePolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateWorkspace
      summary: Update a workspace policy
      tags: [Workspaces]
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspacePolicy"
      responses:
        "200":
          description: Updated workspace policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspacePolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteWorkspace
      summary: Delete a workspace policy
      tags: [Workspaces]
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "204":
          description: Workspace deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/keys:
    get:
      operationId: listApiKeys
      summary: List all API keys (hashed, no raw keys returned)
      tags: [Auth]
      responses:
        "200":
          description: Array of API key records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKeyRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createApiKey
      summary: Generate a new API key
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyBody"
      responses:
        "201":
          description: Newly created API key (raw key shown only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/keys/{id}:
    delete:
      operationId: deleteApiKey
      summary: Revoke an API key
      tags: [Auth]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Key revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/webhooks/{id}:
    delete:
      operationId: deleteWebhook
      summary: Unregister a webhook
      tags: [Webhooks]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Webhook removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        API key passed as a Bearer token. Generate keys via POST /v1/auth/keys.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    InspectRequest:
      type: object
      required: [agentId, framework, action]
      properties:
        agentId:
          type: string
          description: Unique agent identifier
        workspaceId:
          type: string
          description: Optional workspace scope
        framework:
          type: string
          description: Agent framework name (e.g. langchain, autogen)
        protocol:
          $ref: "#/components/schemas/ProtocolKind"
        action:
          $ref: "#/components/schemas/ActionDetail"
        requestedSecrets:
          type: array
          items:
            type: string
          description: Secret names the agent wants injected
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary key-value metadata
...